Massively improve speed of DisplayHost
This commit changes how data is read and written
Notably on the host data is strictly only recieved in 1 thread. The parsing and executing of recieved data/actions is refractored to be multithreaded. Multiple action messages can now be parsed at once on host while preserving the sequentional nature of the actions. This increased the troughput of data that can be recieved by roughly 2000%. This makes it a trivial time cost to send when doing many tests and general debugging.

10/10 would use again for debugging