Unpack dynamic object field type ID from the field
This uncovers a structual problem with the field API. The size of the object ID changes from when it is probed for size and when it's actually written. This causes a safety trip even though it's not in concept an invalid state. Not wise to silence the safety trip in this case. Best to patch the hack with a proper implementation even at the cost of extra complexity.

Need to replace the "return fields to be rewritten when writing a field" with a generator API. Fields will be able to declare a lambda or something that properly states what fields need to be generated and how before the struct is probed in any way.
I'm 96.572% sure that it's gonna be faster and there will be no need to do the buffer reset and rewrite fields hack for all field writing in pipe implementation.