Implement Field Generators API
This states that the process of sizing and writing a field should consist of a new pre processing step if nessecary. A field is able to declare a generator (ValueGeneratorInfo) that declares a generator lambda and a generator target field. Before an object is written but after an IO pool os bound, generators will kick in and generate any fields, but more spesifically, any IO pool virtual fields. This completes the change that removed the ability to notify the pipe if a field should be rewritten.
Now instead of sloppily notifying the pipe about bad fields, they will be selectivelly generated before any rewrites become nessecary.