Refractor MemoryWalker. Introduce concept of saving
MemoryWalker now can handle saving of data it walks over. This is useful because if there is an inline struct inside a struct whos size is modified when walking, it will corrupt the outer struct as it corrupts data in range of previous end of inside struct and its new end.

MemoryWalker handling the saving means it can use its understanding of data it is walking over and delegate the saving of the modification to the most outer struct.

This also makes "user code" simpler. No need to manually use the instance reference and pipe to save the modified data.