Implement grouped NumberSize descriptors as optimization
Before a generic unkown descriptor caused repeated and unneeded access trips to get a number size. Now when 2 dependencies declare they depend on the same number size, all but one of their access trips will be optimized out by relying that all sizes in a group depend on the same size. This means only 1 access folowed by a multiplication by the size of the group has the same effect.

Perfromance improvement is roughly 15% when reading a chunk so extra complexity is worth it as performant reading of a chunk header is important.